feat(chat): replace integer message IDs with Snowflake ID generator

+++ backend/src/SocialTechsy.SocialNetwork.Application/Interfaces/Repositories/IChatRepository.cs
@@ -2,9 +2,6 @@ using SocialTechsy.SocialNetwork.Domain.Entities;
 
 namespace SocialTechsy.SocialNetwork.Application.Interfaces.Repositories;
 
-/// <summary>
-/// Repository interface for Chat entities (Conversation, Message, MessageReaction)
-/// </summary>
 public interface IChatRepository
 {
     // Conversations
@@ -14,7 +11,7 @@ public interface IChatRepository
     Task<Conversation> CreateConversationAsync(Conversation conversation, CancellationToken cancellationToken = default);
     
     // Messages
-    Task<Message?> GetMessageByIdAsync(int messageId, CancellationToken cancellationToken = default);
+    Task<Message?> GetMessageByIdAsync(long messageId, CancellationToken cancellationToken = default);
     Task<(IEnumerable<Message> Items, int TotalCount)> GetMessagesAsync(
         int conversationId,
         int page,
@@ -22,18 +19,18 @@ public interface IChatRepository
         CancellationToken cancellationToken = default);
     Task<IEnumerable<Message>> GetMessagesCursorAsync(
         int conversationId,
-        int? afterMessageId,
+        long? afterMessageId,
         int limit,
         CancellationToken cancellationToken = default);
     Task<Message> AddMessageAsync(Message message, CancellationToken cancellationToken = default);
     Task<bool> RemoveParticipantAsync(int conversationId, int userId, CancellationToken cancellationToken = default);
     Task MarkMessagesAsReadAsync(int conversationId, int userId, CancellationToken cancellationToken = default);
-    Task UpdateDeliveryStatusAsync(int messageId, DeliveryStatus status, CancellationToken cancellationToken = default);
-    Task<IEnumerable<Message>> GetMessagesSinceAsync(int conversationId, int sinceMessageId, int limit = 200, CancellationToken cancellationToken = default);
+    Task UpdateDeliveryStatusAsync(long messageId, DeliveryStatus status, CancellationToken cancellationToken = default);
+    Task<IEnumerable<Message>> GetMessagesSinceAsync(int conversationId, long sinceMessageId, int limit = 200, CancellationToken cancellationToken = default);
     
     // Reactions
-    Task<MessageReaction?> GetReactionAsync(int messageId, int userId, CancellationToken cancellationToken = default);
-    Task<IEnumerable<MessageReaction>> GetMessageReactionsAsync(int messageId, CancellationToken cancellationToken = default);
+    Task<MessageReaction?> GetReactionAsync(long messageId, int userId, CancellationToken cancellationToken = default);
+    Task<IEnumerable<MessageReaction>> GetMessageReactionsAsync(long messageId, CancellationToken cancellationToken = default);
     Task<MessageReaction> AddReactionAsync(MessageReaction reaction, CancellationToken cancellationToken = default);
     Task UpdateReactionAsync(MessageReaction reaction, CancellationToken cancellationToken = default);
     Task RemoveReactionAsync(MessageReaction reaction, CancellationToken cancellationToken = default);
